+++ bindings/nodejs/examples/16b-prepare-output.js
@@ -13,10 +13,15 @@ async function run() {
         const recipientAddress =
             'rms1qrrv7flg6lz5cssvzv2lsdt8c673khad060l4quev6q09tkm9mgtupgf0h0';
         const amount = '1000';
-
+        const tag = '0x12345678'
+        const metadata = '0x00000000025e4b3ca1e3f42320a10700000000000200000001006115000000038a323cf02b18a59e209c66817a057a1000e0b71201006301000000ff0040420f0000000000020000000000'
         const output = await account.prepareOutput({
             recipientAddress,
             amount,
+            features: {
+                tag,
+                metadata
+            }
         });
         const minimumRequiredStorageDeposit =
             await account.minimumRequiredStorageDeposit(output);